extras: Allow superscript to begin with plus, minus, or single quote
This is admittedly a trick, but it is (a) isolated to superscript
parsing, and (b) only applicable to the plus, minus, and single quote
symbols.

The goldmark parser.ScanDelimiter function receives the line, but
doesn't pass it on to anything else. So all we're doing is replacing the
plus, minus, or single quote symbol with the letter "z" if the symbol is
the first character after the delimiter. Using the letter "z" was an
arbitrary choice.

Closes #30